refactor: modularize Markdown chunking strategy with enhanced document parsing and metadata extraction
- Introduced a modular architecture for the Markdown chunking strategy, separating concerns into distinct components: DocumentParser, SectionSplitter, MetadataExtractor, and ChunkProcessor.
- Updated the README to reflect the new release version 0.4.15.
- Removed the old markdown_strategy.py file and replaced it with a more organized structure for better maintainability and readability.
