+++ README.md
@@ -5,7 +5,7 @@
 [![Test Coverage](https://img.shields.io/badge/coverage-view%20reports-blue)](https://qdrant-loader.net/coverage/)
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 
-📋 **[Release Notes v0.4.14](./RELEASE_NOTES.md)** - Latest improvements and bug fixes (July 11, 2025)
+📋 **[Release Notes v0.4.15](./RELEASE_NOTES.md)** - Latest improvements and bug fixes (July 22, 2025)
 
 A comprehensive toolkit for loading data into Qdrant vector database with advanced MCP server support for AI-powered development workflows.
