+++ packages/qdrant-loader/src/qdrant_loader/core/chunking/strategy/markdown/__init__.py
@@ -1,0 +1,5 @@
+"""Markdown chunking strategy package."""
+
+from .markdown_strategy import MarkdownChunkingStrategy
+
+__all__ = ["MarkdownChunkingStrategy"]
